move CBOR parser to Read hierarchy
