+++ lib/core/cardano-wallet-core.cabal
@@ -272,8 +272,6 @@ library
       Cardano.Wallet.Primitive.Types.TokenPolicy
       Cardano.Wallet.Primitive.Types.TokenQuantity
       Cardano.Wallet.Primitive.Types.Tx
-      Cardano.Wallet.Primitive.Types.Tx.CBOR
-      Cardano.Wallet.Primitive.Types.Tx.CBOR.Hash
       Cardano.Wallet.Primitive.Types.Tx.Constraints
       Cardano.Wallet.Primitive.Types.Tx.SealedTx
       Cardano.Wallet.Primitive.Types.Tx.TransactionInfo
@@ -288,6 +286,8 @@ library
       Cardano.Wallet.Transaction
       Cardano.Wallet.Types.Read
       Cardano.Wallet.Types.Read.Tx
+      Cardano.Wallet.Types.Read.Tx.CBOR
+      Cardano.Wallet.Types.Read.Tx.Hash
       Cardano.Wallet.Unsafe
       Cardano.Wallet.Util
       Cardano.Wallet.Version
